+++ R/stat-ma-eq.R
@@ -50,6 +50,8 @@
 #'   the fitted coefficients.
 #' @param coef.keep.zeros logical Keep or drop trailing zeros when formatting
 #'   the fitted coefficients and F-value.
+#' @param decreasing logical It specifies the order of the terms in the
+#'   returned character string; in increasing (default) or decreasing powers.
 #' @param rr.digits,theta.digits,p.digits integer Number of digits after the
 #'   decimal point to use for R^2, theta and P-value in labels. If \code{Inf},
 #'   use exponential notation with three decimal places.
@@ -189,16 +191,26 @@
 #'   stat_ma_line() +
 #'   stat_ma_eq()
 #'
+#' ggplot(my.data, aes(x, y)) +
+#'   geom_point() +
+#'   stat_ma_line() +
+#'   stat_ma_eq(mapping = use_label("eq"))
+#'
+#' ggplot(my.data, aes(x, y)) +
+#'   geom_point() +
+#'   stat_ma_line() +
+#'   stat_ma_eq(mapping = use_label("eq"), decreasing = TRUE)
+#'
 #' # use_label() can assemble and map a combined label
 #' ggplot(my.data, aes(x, y)) +
 #'   geom_point() +
 #'   stat_ma_line(method = "MA") +
-#'   stat_ma_eq(use_label(c("eq", "R2", "P")))
+#'   stat_ma_eq(mapping = use_label(c("eq", "R2", "P")))
 #'
 #' ggplot(my.data, aes(x, y)) +
 #'   geom_point() +
 #'   stat_ma_line(method = "MA") +
-#'   stat_ma_eq(use_label(c("R2", "P", "theta", "method")))
+#'   stat_ma_eq(mapping = use_label(c("R2", "P", "theta", "method")))
 #'
 #' # using ranged major axis regression
 #' ggplot(my.data, aes(x, y)) +
@@ -206,7 +218,7 @@
 #'   stat_ma_line(method = "RMA",
 #'                range.y = "interval",
 #'                range.x = "interval") +
-#'   stat_ma_eq(use_label(c("eq", "R2", "P")),
+#'   stat_ma_eq(mapping = use_label(c("eq", "R2", "P")),
 #'              method = "RMA",
 #'              range.y = "interval",
 #'              range.x = "interval")
@@ -215,7 +227,7 @@
 #' ggplot(my.data, aes(x, y)) +
 #'   geom_point() +
 #'   stat_ma_line(method = "MA") +
-#'   stat_ma_eq(use_label(c("eq", "R2")),
+#'   stat_ma_eq(mapping = use_label(c("eq", "R2")),
 #'              method = "MA",
 #'              nperm = 0)
 #'
@@ -224,13 +236,13 @@
 #'   geom_point() +
 #'   stat_ma_line(formula = x ~ y) +
 #'   stat_ma_eq(formula = x ~ y,
-#'              use_label(c("eq", "R2", "P")))
+#'              mapping = use_label(c("eq", "R2", "P")))
 #'
 #' # modifying both variables within aes()
 #' ggplot(my.data, aes(log(x + 10), log(y + 10))) +
 #'   geom_point() +
 #'   stat_poly_line() +
-#'   stat_poly_eq(use_label("eq"),
+#'   stat_poly_eq(mapping = use_label("eq"),
 #'                eq.x.rhs = "~~log(x+10)",
 #'                eq.with.lhs = "log(y+10)~~`=`~~")
 #'
@@ -245,7 +257,7 @@
 #'        aes(x, y,  shape = group, linetype = group, grp.label = group)) +
 #'   geom_point() +
 #'   stat_ma_line(color = "black") +
-#'   stat_ma_eq(use_label(c("grp", "eq", "R2"))) +
+#'   stat_ma_eq(mapping = use_label(c("grp", "eq", "R2"))) +
 #'   theme_classic()
 #'
 #' # Inspecting the returned data using geom_debug()
@@ -267,7 +279,7 @@
 #' if (gginnards.installed)
 #'   ggplot(my.data, aes(x, y)) +
 #'     geom_point() +
-#'     stat_ma_eq(aes(label = after_stat(eq.label)),
+#'     stat_ma_eq(mapping = aes(label = after_stat(eq.label)),
 #'                geom = "debug",
 #'                output.type = "markdown")
 #'
@@ -301,6 +313,7 @@ stat_ma_eq <- function(mapping = NULL, data = NULL,
                        small.p = FALSE,
                        coef.digits = 3,
                        coef.keep.zeros = TRUE,
+                       decreasing = FALSE,
                        rr.digits = 2,
                        theta.digits = 2,
                        p.digits = max(1, ceiling(log10(nperm))),
@@ -377,6 +390,7 @@ stat_ma_eq <- function(mapping = NULL, data = NULL,
                    small.p = small.p,
                    coef.digits = coef.digits,
                    coef.keep.zeros = coef.keep.zeros,
+                   decreasing = decreasing,
                    rr.digits = rr.digits,
                    theta.digits = theta.digits,
                    p.digits = p.digits,
@@ -419,6 +433,7 @@ ma_eq_compute_group_fun <- function(data,
                                     small.p,
                                     coef.digits,
                                     coef.keep.zeros,
+                                    decreasing,
                                     rr.digits,
                                     theta.digits,
                                     p.digits,
@@ -618,6 +633,7 @@ ma_eq_compute_group_fun <- function(data,
     eq.char <- coefs2poly_eq(coefs = coefs,
                              coef.digits = coef.digits,
                              coef.keep.zeros = coef.keep.zeros,
+                             decreasing = decreasing,
                              eq.x.rhs = eq.x.rhs,
                              lhs = lhs,
                              output.type = output.type,